+++ core/web/resolver/mutation.go
@@ -166,6 +166,7 @@ type createFeedsManagerChainConfigInput struct {
 	OCR2Multiaddr      *string
 	OCR2P2PPeerID      *string
 	OCR2KeyBundleID    *string
+	OCR2Plugins        string
 }
 
 func (r *Resolver) CreateFeedsManagerChainConfig(ctx context.Context, args struct {
@@ -209,12 +210,18 @@ func (r *Resolver) CreateFeedsManagerChainConfig(ctx context.Context, args struc
 	}
 
 	if args.Input.OCR2Enabled {
+		var plugins feeds.Plugins
+		if err = plugins.Scan(args.Input.OCR2Plugins); err != nil {
+			return nil, err
+		}
+
 		params.OCR2Config = feeds.OCR2Config{
 			Enabled:     args.Input.OCR2Enabled,
 			IsBootstrap: *args.Input.OCR2IsBootstrap,
 			Multiaddr:   null.StringFromPtr(args.Input.OCR2Multiaddr),
 			P2PPeerID:   null.StringFromPtr(args.Input.OCR2P2PPeerID),
 			KeyBundleID: null.StringFromPtr(args.Input.OCR2KeyBundleID),
+			Plugins:     plugins,
 		}
 	}
 
@@ -292,6 +299,7 @@ type updateFeedsManagerChainConfigInput struct {
 	OCR2Multiaddr      *string
 	OCR2P2PPeerID      *string
 	OCR2KeyBundleID    *string
+	OCR2Plugins        string
 }
 
 func (r *Resolver) UpdateFeedsManagerChainConfig(ctx context.Context, args struct {
@@ -329,12 +337,18 @@ func (r *Resolver) UpdateFeedsManagerChainConfig(ctx context.Context, args struc
 	}
 
 	if args.Input.OCR2Enabled {
+		var plugins feeds.Plugins
+		if err = plugins.Scan(args.Input.OCR2Plugins); err != nil {
+			return nil, err
+		}
+
 		params.OCR2Config = feeds.OCR2Config{
 			Enabled:     args.Input.OCR2Enabled,
 			IsBootstrap: *args.Input.OCR2IsBootstrap,
 			Multiaddr:   null.StringFromPtr(args.Input.OCR2Multiaddr),
 			P2PPeerID:   null.StringFromPtr(args.Input.OCR2P2PPeerID),
 			KeyBundleID: null.StringFromPtr(args.Input.OCR2KeyBundleID),
+			Plugins:     plugins,
 		}
 	}