+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- TOML configuration and secrets are now scoped to `chainlink node start` command rather than being global flags
+- TOML configuration and secrets are now scoped to `chainlink node` command rather than being global flags.
 
 ### Removed